resolved comment's on PR and add core-component for Row Col
