+++ react-frontend/src/core-components/button/ButtonComponent.js~HEAD
@@ -1,0 +1,1 @@
+export { Button } from "react-bootstrap";
