+++ react-frontend/src/core-components/button/ButtonComponent.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import ButtonComponent from "core-components/Button/ButtonComponent";
+import ButtonComponent from "core-components/button/ButtonComponent";
 
 it("ButtonComponent should equal to submit text", () => {
   const { getByText } = render(<ButtonComponent text="submit" type="submit" />);
