+++ react-frontend/src/core-components/col/ColComponent.js
@@ -1,0 +1,3 @@
+import { Col } from "react-bootstrap";
+
+export default Col;
