+++ react-frontend/src/core-components/label/LabelComponent.js
