+++ react-frontend/src/core-components/label/LabelComponent.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import LabelComponent from "core-components/Label/LabelComponent";
+import LabelComponent from "core-components/label/LabelComponent";
 
 it("LabelComponent should equal to submit text", () => {
   const { getByText } = render(<LabelComponent text="submit" />);
