+++ react-frontend/src/core-components/row/RowComponent.js
@@ -1,0 +1,3 @@
+import { Row } from "react-bootstrap";
+
+export default Row;
